(Malak) Final touches
just changed the function "calculateCost" attribute in the Booking class back to int and instead type casted the "(int) diffInDays" in the Receptionist class

+++ src/houserental/Booking.java
@@ -107,7 +107,7 @@ public class Booking {
         this.totalCost = totalCost;
     }
 
- public double calculateCost(long numberOfNights) { //changed from int to long
+ public double calculateCost(int numberOfNights) { //changed from int to long
       
         return numberOfNights * costPerNight;  ///Definition of `costPerNight`
